config: simplify presale input
